Try to import missing archives from incoming
If you run "0repo update" and some archives are missing, check if
they're in the incoming directory. If so, try to import them and
continue. If not, tell the user to put them there.

This makes it easier to add missing archives and have them uploaded in
the usual way, and with their digests checked.